# Use managed memory for NDSH benchmarks