@karthikeyann 2024-10-11
The benchmarked query did not have any effect due to this PR change. (tested with Q05) 

This is overall runtime for single run with data generation. eg. `./benchmark --axis scale_factor=20 --run-once`
Old = This PR.
New = new nvbench fixture which delays creation of rmm_mode after data generation. Data generation uses managed pool.
[new_ndsh_fixture.patch](https://github.com/user-attachments/files/17346686/new_ndsh_fixture.patch)

| Benchmark             | Old Time SF=10  | New Time SF=10   | Old Time SF=20  | New Time SF=20  |
|-----------------------|-----------------|------------------|-----------------|------------------|
| NDSH_Q01_NVBENCH      | 0m10.998s       | 0m10.798s        | 0m31.818s       | 0m18.388s        |
| NDSH_Q05_NVBENCH      | 0m13.232s       | 0m12.652s        | 0m35.439s       | 0m22.404s        |
| NDSH_Q06_NVBENCH      | 0m11.227s       | 0m11.009s        | 0m31.991s       | 0m18.114s        |
| NDSH_Q09_NVBENCH      | 0m15.518s       | 0m14.674s        | 0m39.800s       | 0m26.957s        |
| NDSH_Q10_NVBENCH      | 0m13.164s       | 0m12.669s        | 0m35.183s       | 0m22.056s        |

If this data generation time does not matter, we can merge this PR.