@karthikeyann 2024-10-19
@GregoryKimball 

| TESTED VARIANTS                                  | Q05 Max SF |
| ------------------------------------------------ | ---------- |
| branch-24.12 (async)                             | 16         |
| host side staging (async)                        | 18         |
| hss + managed pool for datagen, (async)          | 30         |
| hss + chunked pq (async)                         | 40         |
| hss + chunked pq  (rmm pool)                     | 27         |
| hss + chunked pq + datagen=managed_pool, (async or pool) | 60         |

The final version in this PR is hss + chunked pq + datagen=managed_pool.