@GregoryKimball 2024-10-21
Thank you @karthikeyann for studying this. It looks like the chunked PQ writer has a big impact as well - thank you for identifying that. 

I'm happy to proceed with the current state of `hss + chunked pq + datagen=managed_pool`, although `hss + chunked pq` is acceptable as well.